requirements.txt
Flask
opentelemetry-distro
opentelemetry-exporter-otlp